Build a fcgi and non-fcgi version of the library
The simple FastCGI protocol, fcgi_stdio, redefines a lot of the libc
stdio functions, which trips up libctemplate. Hence we included the
fcgi_stio.h header file in ctemplate.c so that libctemplate used the
fcgi versions.

However there may be times when you want to use the libc versions
instead, e.g when not using FastCGI.

The simple thing to do is just build two versions of the library, a
normal non-fcgi version and one that uses the fcgi functions.

Hence we now have the following

    libctemplate
    libctemplate-fcgi

and applications can choose to link against which version they want.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>