Build a version for the FastCGI FCGX_ API
The FCGX_ API as defined in fcgiapp.h is the low level FastCGI API upon
which the FCGI_ (fcgi_stdio.h) API is built on.

FCGX has some advantages, such as not wrapping the stdio library.

I introduced some #if'deffery to switch between a FILE * and a
FCGX_Stream * depending on if we are building the FCGX version or not.

I converted a couple of fputs() to fprintf() so that the only calls to
fputs are those which are writing out the template (possibly to a
FCGX_Stream *). As such, if we are building for FCGX (by defining
_HAVE_FCGX) then we remap fputs() and fputc() to their FCGX_
equivalents.

This library is called

    libctemplate-fcgx

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>